+++ lib/src/models/editor_options.dart
@@ -1,4 +1,4 @@
-import 'package:dart_helper_utils/dart_helper_utils.dart';
+import 'package:flutter_helper_utils/flutter_helper_utils.dart';
 import 'package:flutter_monaco/flutter_monaco.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';